Removes cv.html
Converting the cv to a html was a hack but did not work because the
three dashes jekyll uses to parse files are removed. I can not be
expected to fix this every time I run the haml compiler.

The solution is to host the website on my own web server