Makes change in summary
The current summary talks about how I add the same value as someone with
at least 5 years experience but it doesn't give a clue of how this skill
was achieved.

Point out that I have been working on a social network for the past 6
years and it's it the reason I know so much. It keeps me researching to
build it in a better way. Since I have to keep it short, just point out
that I am more than willing to chat about it more into detail.